requirements.txt
# Core dependencies for NYC Services GPT RAG system
flask>=2.3.0
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
google-generativeai>=0.3.0
pandas>=2.0.0
numpy>=1.24.0

# Vector database and embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Data processing
requests>=2.31.0
beautifulsoup4>=4.12.0

# Testing and evaluation
pytest>=7.4.0
pytest-cov>=4.1.0

# Development tools
python-dotenv>=1.0.0
black>=23.0.0
flake8>=6.0.0 